+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 * `TestTrait::deprecated()` is deprecated and will be removed in a future release. Use instead the PHPUnit Bridge;
 * `array_clean()`, `array_has_only_numeric_keys()`, `array_to_string()`, `array_unique_recursive()`, `is_html()`,
   `is_stringable()` and `uncamelcase()` global functions are deprecated and will be removed in a future release;
+* `clean_url()`, `get_hostname_from_url()`, `is_external_url()` and `is_localhost()` global network functions are
+  deprecated and will be removed in a future release;
 * removed `cakephp/core` package;
 
 ### 1.9.3